fix(ui): address session delete review feedback
