+++ docs/plans/session-tab-delete-feedback/plan.md
@@ -1,6 +1,7 @@
 ---
 spec: docs/specs/ui/session-tab-delete-feedback.md
 created: 2026-08-05
+amended: 2026-08-22
 status: shipped
 ---
 
@@ -12,7 +13,9 @@ Give the desktop Dockview session-tab close flow local pending feedback while pr
 session lifecycle behavior used by context menus and mobile. First make the shared delete action
 selectively suppress progress/success toasts, then replace Dockview's opaque close icon with a
 small repository-owned action that can render an X or spinner. Finish with focused desktop and
-mobile Playwright coverage.
+mobile Playwright coverage. A later localized-confirmation refinement keeps the tab-X dialog,
+anchors context-menu confirmation to its Delete item, and morphs the phone picker row into inline
+touch actions.
 
 ## Frontend
 
@@ -44,18 +47,23 @@ Add the close action's accessible label to `apps/web/src/locales/en/common.json`
 
 ### Mobile design contract
 
-- **Desktop outcome:** the session-tab X is the progress surface and remains in the existing tab
-  strip; no overlay or layout change is introduced.
+- **Desktop outcome:** the session-tab X remains the progress surface and retains its alert dialog.
+  Context-menu Delete keeps the menu mounted and opens a compact confirmation popover anchored to
+  that item.
 - **Mobile entry point:** the existing `MobileSessionsPicker` pill opens `MobilePickerSheet`, and a
-  session row's visible actions menu owns deletion. There is no Dockview X on phone.
+  session row's visible actions menu owns deletion. The target row then morphs into touch-sized
+  inline Cancel and Delete actions; there is no Dockview X or nested confirmation dialog on phone.
 - **Nearest shipped exemplar:** `apps/web/components/task/mobile/mobile-sessions-section.tsx`
   remains authoritative for the phone hierarchy, bottom-sheet surface, internal scroll owner,
   safe-area handling, and touch targets.
-- **Shared versus specialized behavior:** the delete transport/store mutation remains shared in
-  `useSessionActions`; only the desktop tab-X presentation and feedback mode specialize. Mobile
-  markup and its default feedback mode do not change.
-- **Parity proof:** a mobile Playwright scenario deletes one of two sessions through the Sessions
-  picker and verifies the selected session disappears while the remaining session stays reachable.
+- **Shared versus specialized behavior:** the delete transport/store mutation and localized warning
+  copy remain shared. Desktop context-menu and phone picker components specialize only the local
+  confirmation presentation; their default request feedback remains unchanged.
+- **Dismissal:** Cancel, session selection, and externally closing the picker clear pending inline
+  confirmation without dispatching deletion.
+- **Parity proof:** desktop Playwright scenarios cover the anchored context-menu confirmation. A
+  mobile scenario cancels and then confirms the inline row action, proving the selected session
+  disappears while the remaining session stays reachable.
 
 ## Tests
 
@@ -67,6 +75,9 @@ Add the close action's accessible label to `apps/web/src/locales/en/common.json`
   `apps/web/components/task/session-tab-close-action.test.tsx` to prove the idle X is operable, the
   pending state renders a status spinner with `aria-busy`, and pending activation cannot dispatch a
   second callback.
+- **Localized confirmation surfaces:** `session-tab-menu.test.tsx` proves the context menu remains
+  mounted for its anchored popover. `mobile-sessions-section.test.tsx` proves inline confirm,
+  cancellation, dispatch, touch-target sizing, and cleanup when the picker closes externally.
 
 ## E2E Tests
 
@@ -74,9 +85,14 @@ Add the close action's accessible label to `apps/web/src/locales/en/common.json`
   THEN the tab/session is removed and no `Deleting session...` or successful-deletion toast appears.
   **File:** `apps/web/e2e/tests/session/session-tab-management.spec.ts`.
 - **Scenario:** GIVEN two deletable sessions on a phone viewport, WHEN the user opens the Sessions
-  picker and deletes one through its row action, THEN that row disappears and the remaining session
-  stays reachable. **File:** `apps/web/e2e/tests/session/mobile-session-deletion.spec.ts`, run by
-  the `mobile-chrome` project.
+  picker and chooses Delete, THEN the row shows inline confirmation with no alert dialog; cancelling
+  preserves the row, and confirming removes it while the remaining session stays reachable.
+  **File:** `apps/web/e2e/tests/session/mobile-session-deletion.spec.ts`, run by the `mobile-chrome`
+  project.
+- **Scenario:** GIVEN a desktop session context menu, WHEN the user chooses Delete, THEN a compact
+  anchored popover appears without an alert dialog and the existing cancel/confirm deletion outcomes
+  remain intact. **Files:** `apps/web/e2e/tests/session/multi-session-ux.spec.ts` and
+  `apps/web/e2e/tests/session/session-tab-management.spec.ts`.
 
 The transient spinner is covered deterministically by the component test because the real delete
 response can settle too quickly for a race-free Playwright observation. Playwright covers the
@@ -101,6 +117,18 @@ The pseudo-locale artifact updated is `apps/web/src/locales/pseudo/common.json`.
 used isolated temporary backends and cleaned their generated results; no external systems were
 changed.
 
+### Review remediation (2026-08-22)
+
+- `pnpm --filter @kandev/web test -- components/task/mobile/mobile-sessions-section.test.tsx components/task/session-tab-menu.test.tsx` - 2 files, 17 tests passed.
+- `pnpm run typecheck` - passed.
+- Targeted ESLint for the shared description, desktop menu, mobile picker, and regression test -
+  passed with no warnings.
+- `pnpm run i18n:check` and `pnpm run i18n:ratchet` - passed; catalogs remain complete and four
+  modified frontend files passed the new-code ratchet.
+- Local Playwright was not repeated because remediation changes only test coverage, comments,
+  component ownership, and internal documentation; exact-head CI reruns the integrated desktop and
+  mobile scenarios.
+
 ## Implementation Waves And Parallel Candidates
 
 Wave 1:
@@ -113,6 +141,11 @@ Wave 2:
 
 Execution is sequential in the primary conversation. The E2E task depends on the production and
 unit-test changes from task 01; these tasks are not parallel-safe.
+
+## Public documentation
+
+No public documentation change. This refines transient confirmation presentation without changing
+commands, configuration, navigation terminology, or a public API.
 
 ## Risks
